pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "brain-mri-triage"
version = "0.1.0"
description = "FastAPI app para triagem complementar de MRI cerebral com modelo treinado."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "tensorflow==2.18.0",
    "numpy==2.0.2",
    "pandas==3.0.2",
    "scikit-learn==1.8.0",
    "matplotlib==3.10.9",
    "opencv-python-headless==4.13.0.92",
    "Pillow==12.2.0",
    "fastapi==0.136.1",
    "uvicorn==0.46.0",
    "python-multipart==0.0.27",
]

[tool.setuptools.packages.find]
where = ["src"]
